pyproject.toml
[project]
name = "drag"
version = "0.1.0"
description = "DRAG: Document Retrieval with Agentic Grounding - Finetuning VLMs for better sparse search"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "JVL" }
]
keywords = ["vlm", "retrieval", "finetuning", "rag", "document-qa"]

dependencies = [
    # Core
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "tqdm>=4.65.0",
    "datasets>=2.14.0",
    # Search
    "whoosh>=2.7.4",
    # Image handling
    "pillow>=10.0.0",
    "pdf2image>=1.16.0",
    "pdfplumber>=0.11.0",
    # Finetuning (Qwen3-VL requires transformers>=4.57.0)
    "torch>=2.0.0",
    "transformers>=4.57.0",
    "peft>=0.10.0",
    "trl>=0.8.0",
    "bitsandbytes>=0.43.0",
    "accelerate>=0.30.0",
    "torchvision>=0.18.0",
    # Qwen3-VL specific
    "qwen-vl-utils>=0.0.14",
    # Evaluation
    "anls-star>=0.1.0",
    # Logging
    "tensorboardx>=2.6.0",
    "matplotlib>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
vllm = [
    "vllm>=0.4.0",
]

[project.urls]
Repository = "https://github.com/your-username/DRAG"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agentic-retrieval-finetuning"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

